refactor(batching): move the batching code into a layer on the HTTP service
This is _essentially_ just a fancy cut + paste. But the new `JoinBatchRequestsLayer` applies _before_ plugins at the `http_client_service` hook point.

The `JoinBatchRequestsLayer` still ends up using the `do_fetch` function
from the `SubgraphService`. The functionality of the `do_fetch` function
should be moved into a layer on the `HttpClientService`.
