+++ apollo-router/src/services/subgraph/service.rs
@@ -8,30 +8,19 @@ use std::task::Poll;
 use futures::future::BoxFuture;
 use http::StatusCode;
 use http_body::Body as _;
-use itertools::Itertools;
 use opentelemetry::Key;
 use opentelemetry::KeyValue;
-use tokio::sync::oneshot;
 use tower::BoxError;
 use tower::ServiceBuilder;
 use tower::ServiceExt;
 use tracing::Instrument;
-use tracing::instrument;
 
-use super::SubgraphRequestId;
 use super::http::do_fetch;
 use super::http::get_uri_details;
 use super::http::http_response_to_graphql_response;
-use crate::Context;
 use crate::Notify;
-use crate::batching::BatchQueryInfo;
-use crate::batching::SubgraphBatchRequest;
-use crate::batching::assemble_batch;
-use crate::configuration::BatchingMode;
 use crate::configuration::SubgraphApq;
 use crate::configuration::subgraph::SubgraphConfiguration;
-use crate::error::FetchError;
-use crate::error::SubgraphBatchingError;
 use crate::graphql;
 use crate::json_ext::Object;
 use crate::layers::InternalServiceBuilderExt as _;
@@ -48,13 +37,10 @@ use crate::plugins::telemetry::consts::SUBGRAPH_REQUEST_SPAN_NAME;
 use crate::services::Plugins;
 use crate::services::SubgraphRequest;
 use crate::services::SubgraphResponse;
-use crate::services::http::HttpResponse;
 use crate::services::http::service::WireByteCount;
 use crate::services::layers::apq::subgraph::SubgraphApqLayer;
-use crate::services::layers::content_negotiation::ContentType;
 use crate::services::layers::content_negotiation::SubgraphContentNegotiationLayer;
 use crate::services::router;
-use crate::services::router::body::RouterBody;
 use crate::services::subgraph;
 
 /// Client for interacting with subgraphs.
@@ -96,357 +82,6 @@ impl tower::Service<SubgraphRequest> for SubgraphService {
 
         Box::pin(async move { call_http(request, http_client, &service_name).await })
     }
-}
-
-/// Process a single subgraph batch request
-#[instrument(skip(http_client, contexts, request))]
-pub(crate) async fn process_batch(
-    http_client: crate::services::http::BoxCloneService,
-    service: &str,
-    mut contexts: Vec<(Context, SubgraphRequestId)>,
-    request: http::Request<RouterBody>,
-    listener_count: usize,
-) -> Result<Vec<HttpResponse>, FetchError> {
-    // The graphql spec is lax about what strategy to use for processing responses: https://github.com/graphql/graphql-over-http/blob/main/spec/GraphQLOverHTTP.md#processing-the-response
-    //
-    // "If the response uses a non-200 status code and the media type of the response payload is application/json
-    // then the client MUST NOT rely on the body to be a well-formed GraphQL response since the source of the response
-    // may not be the server but instead some intermediary such as API gateways, proxies, firewalls, etc."
-    //
-    // The TLDR of this is that it's really asking us to do the best we can with whatever information we have with some modifications depending on content type.
-    // Our goal is to give the user the most relevant information possible in the response errors
-    //
-    // Rules:
-    // 1. If the content type of the response is not `application/json` or `application/graphql-response+json` then we won't try to parse.
-    // 2. If an HTTP status is not 2xx it will always be attached as a graphql error.
-    // 3. If the response type is `application/json` and status is not 2xx and the body the entire body will be output if the response is not valid graphql.
-
-    // We need a "representative context" for a batch. We use the first context in our list of
-    // contexts
-    let batch_context = contexts
-        .first()
-        .expect("we have at least one context in the batch")
-        .0
-        .clone();
-    let service_name = service.to_string();
-
-    // Update our batching metrics (just before we fetch)
-    u64_histogram!(
-        "apollo.router.operations.batching.size",
-        "Number of queries contained within each query batch",
-        listener_count as u64,
-        mode = BatchingMode::BatchHttpLink.to_string(), // Only supported mode right now
-        subgraph = service_name.clone()
-    );
-
-    u64_counter!(
-        "apollo.router.operations.batching",
-        "Total requests with batched operations",
-        1,
-        // XXX(@goto-bus-stop): Should these be `batching.mode`, `batching.subgraph`?
-        // Also, other metrics use a different convention to report the subgraph name
-        mode = BatchingMode::BatchHttpLink.to_string(), // Only supported mode right now
-        subgraph = service_name.clone()
-    );
-
-    // Perform the actual fetch. If this fails then we didn't manage to make the call at all, so we can't do anything with it.
-    tracing::debug!("fetching from subgraph: {service}");
-    let (parts, content_type, body) =
-        match do_fetch(http_client, &batch_context, service, request).await {
-            Ok(res) => res,
-            Err(err) => {
-                let resp = http::Response::builder()
-                    .status(StatusCode::INTERNAL_SERVER_ERROR)
-                    .body(err.to_graphql_error(None))
-                    .map_err(|err| FetchError::SubrequestHttpError {
-                        status_code: None,
-                        service: service_name.clone(),
-                        reason: format!("cannot create the http response from error: {err:?}"),
-                    })?;
-                let (parts, body) = resp.into_parts();
-                let body =
-                    serde_json::to_vec(&body).map_err(|err| FetchError::SubrequestHttpError {
-                        status_code: None,
-                        service: service_name.clone(),
-                        reason: format!("cannot serialize the error: {err:?}"),
-                    })?;
-                (
-                    parts,
-                    Ok(ContentType::ApplicationJson),
-                    Some(Ok(body.into())),
-                )
-            }
-        };
-
-    // Mask sensitive response headers once, for reuse in both the telemetry
-    // event and the debug log below. Logging the raw `parts` would otherwise
-    // leak the very header values this masking redacts.
-    let headers_str = crate::services::header_masking::masked_headers_for_log(
-        &batch_context,
-        crate::services::header_masking::Direction::Response,
-        Some(service),
-        &parts.headers,
-    );
-
-    tracing::debug!(
-        "parts status: {:?}, version: {:?}, headers: {headers_str}, content_type: {content_type:?}, body: {body:?}",
-        parts.status,
-        parts.version,
-    );
-    let value =
-        serde_json::from_slice(&body.ok_or(FetchError::SubrequestMalformedResponse {
-            service: service_name.clone(),
-            reason: "no body in response".to_string(),
-        })??)
-        .map_err(|error| FetchError::SubrequestMalformedResponse {
-            service: service_name.clone(),
-            reason: error.to_string(),
-        })?;
-
-    tracing::debug!("json value from body is: {value:?}");
-
-    let array = ensure_array!(value).map_err(|error| FetchError::SubrequestMalformedResponse {
-        service: service_name.clone(),
-        reason: error.to_string(),
-    })?;
-    let mut exploded_bodies = Vec::with_capacity(array.len());
-    for value in array {
-        let object =
-            ensure_object!(value).map_err(|error| FetchError::SubrequestMalformedResponse {
-                service: service_name.clone(),
-                reason: error.to_string(),
-            })?;
-
-        // Map our Vec<u8> into Bytes
-        // Map our serde conversion error to a FetchError
-        let body = serde_json::to_vec(&object).map_err(|error| {
-            FetchError::SubrequestMalformedResponse {
-                service: service_name.clone(),
-                reason: error.to_string(),
-            }
-        })?;
-
-        exploded_bodies.push(router::body::from_bytes(body));
-    }
-
-    tracing::debug!("we have a vec of graphql_responses: {exploded_bodies:?}");
-    // Before we process our graphql responses, ensure that we have a context for each
-    // response
-    if exploded_bodies.len() != contexts.len() {
-        return Err(FetchError::SubrequestBatchingError {
-            service: service_name.clone(),
-            reason: format!(
-                "number of contexts ({}) is not equal to number of graphql responses ({})",
-                contexts.len(),
-                exploded_bodies.len()
-            ),
-        });
-    }
-
-    // We are going to pop contexts from the back, so let's reverse our contexts
-    contexts.reverse();
-    // Build an http Response for each graphql response
-    let exploded_responses: Result<Vec<_>, _> = exploded_bodies
-        .into_iter()
-        .map(|body| {
-            http::Response::builder()
-                .status(parts.status)
-                .version(parts.version)
-                .body(body)
-                .map(|mut http_response| {
-                    *http_response.headers_mut() = parts.headers.clone();
-                    // Use the original context for the request to create the response
-                    let (context, _id) =
-                        contexts.pop().expect("we have a context for each response");
-                    let resp = HttpResponse {
-                        http_response,
-                        context,
-                    };
-
-                    // Avoid `{resp:?}`: SubgraphResponse's derived Debug prints
-                    // the response HeaderMap unmasked. Log the non-header parts.
-                    tracing::debug!(
-                        "built subgraph response for {service}: status={:?}, body={:?}",
-                        resp.http_response.status(),
-                        resp.http_response.body(),
-                    );
-                    resp
-                })
-                .map_err(|e| FetchError::MalformedResponse {
-                    reason: e.to_string(),
-                })
-        })
-        .collect();
-
-    match &exploded_responses {
-        Ok(responses) => tracing::debug!("built {} subgraph responses", responses.len()),
-        Err(error) => tracing::debug!("failed to build subgraph responses: {error}"),
-    }
-    exploded_responses
-}
-
-/// Notify all listeners of a batch query of the results
-async fn notify_batch_query(
-    service: String,
-    senders: Vec<oneshot::Sender<Result<HttpResponse, BoxError>>>,
-    responses: Result<Vec<HttpResponse>, FetchError>,
-) -> Result<(), BoxError> {
-    // Avoid `{responses:#?}`: SubgraphResponse's derived Debug prints the
-    // response HeaderMap unmasked. Log the listener count and a result summary.
-    match &responses {
-        Ok(responses) => tracing::debug!(
-            "handling response for service '{service}' with {} listeners: {} responses",
-            senders.len(),
-            responses.len(),
-        ),
-        Err(error) => tracing::debug!(
-            "handling response for service '{service}' with {} listeners: error: {error}",
-            senders.len(),
-        ),
-    }
-
-    match responses {
-        // If we had an error processing the batch, then pipe that error to all of the listeners
-        Err(e) => {
-            for tx in senders {
-                // Try to notify all waiters. If we can't notify an individual sender, then log an error
-                // which, unlike failing to notify on success (see below), contains the the entire error
-                // response.
-                if let Err(log_error) = tx.send(Err(Box::new(e.clone()))).map_err(|_| {
-                    FetchError::SubrequestBatchingError {
-                        service: service.clone(),
-                        reason: format!("tx send failed: {e:?}"),
-                    }
-                }) {
-                    tracing::error!(service, error=%log_error, "failed to notify sender that batch processing failed");
-                }
-            }
-        }
-
-        Ok(rs) => {
-            // Before we process our graphql responses, ensure that we have a tx for each
-            // response
-            if senders.len() != rs.len() {
-                return Err(Box::new(FetchError::SubrequestBatchingError {
-                    service,
-                    reason: format!(
-                        "number of txs ({}) is not equal to number of graphql responses ({})",
-                        senders.len(),
-                        rs.len()
-                    ),
-                }));
-            }
-
-            // We have checked before we started looping that we had a tx for every
-            // graphql_response, so zip_eq shouldn't panic.
-            // Use the tx to send a graphql_response message to each waiter.
-            for (response, sender) in rs.into_iter().zip_eq(senders) {
-                if let Err(log_error) = sender
-                    .send(Ok(response))
-                    // If we fail to notify the waiter that our request succeeded, do not log
-                    // out the entire response since this may be substantial and/or contain
-                    // PII data. Simply log that the send failed.
-                    .map_err(|_error| FetchError::SubrequestBatchingError {
-                        service: service.to_string(),
-                        reason: "tx send failed".to_string(),
-                    })
-                {
-                    tracing::error!(service, error=%log_error, "failed to notify sender that batch processing succeeded");
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-struct BatchInfo {
-    service: String,
-    /// A pre-readied HTTP client service for this subgraph.
-    http_client: crate::services::http::BoxCloneService,
-    request: http::Request<RouterBody>,
-    contexts: Vec<(Context, SubgraphRequestId)>,
-}
-
-type BatchResult = (
-    BatchInfo,
-    Vec<oneshot::Sender<Result<HttpResponse, BoxError>>>,
-);
-
-/// Collect all batch requests and process them concurrently
-///
-/// # Panics
-/// The HTTP client services inside the svc_map must already be readied: otherwise, it may panic.
-#[instrument(skip_all)]
-pub(crate) async fn process_batches(
-    svc_map: HashMap<String, Vec<BatchQueryInfo>>,
-) -> Result<(), BoxError> {
-    // We need to strip out the senders so that we can work with them separately.
-    let mut errors = vec![];
-    let (info, txs): (Vec<_>, Vec<_>) =
-        futures::future::join_all(svc_map.into_iter().map(|(service, requests)| async {
-            let SubgraphBatchRequest {
-                http_client,
-                contexts,
-                request,
-                txs,
-            } = assemble_batch(requests).await?;
-
-            Ok((
-                BatchInfo {
-                    service,
-                    http_client,
-                    request,
-                    contexts,
-                },
-                txs,
-            ))
-        }))
-        .await
-        .into_iter()
-        .filter_map(|x: Result<BatchResult, BoxError>| x.map_err(|e| errors.push(e)).ok())
-        .unzip();
-
-    // If errors isn't empty, then process_batches cannot proceed. Let's log out the errors and
-    // return
-    if !errors.is_empty() {
-        for error in errors {
-            tracing::error!("assembling batch failed: {error}");
-        }
-        return Err(SubgraphBatchingError::ProcessingFailed(
-            "assembling batches failed".to_string(),
-        )
-        .into());
-    }
-    // It is not ok to panic if the length of the txs and info do not match. Let's make sure they
-    // do
-    if txs.len() != info.len() {
-        return Err(SubgraphBatchingError::ProcessingFailed(
-            "length of txs and info are not equal".to_string(),
-        )
-        .into());
-    }
-    let batch_futures = info.into_iter().zip_eq(txs).map(
-        |(
-            BatchInfo {
-                service,
-                http_client,
-                request,
-                contexts,
-            },
-            senders,
-        )| async move {
-            let listener_count = senders.len();
-            let batch_result =
-                process_batch(http_client, &service, contexts, request, listener_count).await;
-
-            notify_batch_query(service, senders, batch_result).await
-        },
-    );
-
-    futures::future::try_join_all(batch_futures).await?;
-
-    Ok(())
 }
 
 /// call_http makes http calls with modified graphql::Request (body)
